fix(solver): cfl edge case
Previously, if adaptive_timestep is set to true while the solver is
instantiated and then CFL is specified as a command line argument, that
CFL is respected only if `--adaptive-timestep` ccommand line argument is
also passed alongwith.

Now if adaptive_timestep is set to true while instantiating the solver
and CFL is specified as a command line argument, that CFL is used.